refactor: Canvas, Room 컴포넌트 분리
